Moving agent from participant2 to participant
