+++ src/main/scala/edu/ie3/simona/model/participant/evcs/EvcsModel.scala
@@ -13,8 +13,8 @@ import edu.ie3.datamodel.models.result.system.{
   EvcsResult,
   SystemParticipantResult,
 }
-import edu.ie3.simona.agent.participant2.ParticipantAgent
-import edu.ie3.simona.agent.participant2.ParticipantAgent.ParticipantRequest
+import edu.ie3.simona.agent.participant.ParticipantAgent
+import edu.ie3.simona.agent.participant.ParticipantAgent.ParticipantRequest
 import edu.ie3.simona.config.RuntimeConfig.EvcsRuntimeConfig
 import edu.ie3.simona.model.participant.ParticipantModel.{
   ModelState,
