+++ src/main/scala/edu/ie3/simona/ontology/messages/services/EvMessage.scala
@@ -6,7 +6,7 @@
 
 package edu.ie3.simona.ontology.messages.services
 
-import edu.ie3.simona.agent.participant2.ParticipantAgent.ParticipantRequest
+import edu.ie3.simona.agent.participant.ParticipantAgent.ParticipantRequest
 import edu.ie3.simona.model.participant.evcs.EvModelWrapper
 import edu.ie3.simona.ontology.messages.services.ServiceMessage.ServiceResponseMessage
 import edu.ie3.simona.service.Data.SecondaryData
