+++ src/main/scala/edu/ie3/simona/service/ev/ExtEvDataService.scala
@@ -6,8 +6,8 @@
 
 package edu.ie3.simona.service.ev
 
-import edu.ie3.simona.agent.participant2.ParticipantAgent
-import edu.ie3.simona.agent.participant2.ParticipantAgent.{
+import edu.ie3.simona.agent.participant.ParticipantAgent
+import edu.ie3.simona.agent.participant.ParticipantAgent.{
   DataProvision,
   RegistrationSuccessfulMessage,
 }
