+++ src/main/scala/edu/ie3/simona/service/weather/WeatherService.scala
@@ -6,8 +6,8 @@
 
 package edu.ie3.simona.service.weather
 
-import edu.ie3.simona.agent.participant2.ParticipantAgent
-import edu.ie3.simona.agent.participant2.ParticipantAgent.{
+import edu.ie3.simona.agent.participant.ParticipantAgent
+import edu.ie3.simona.agent.participant.ParticipantAgent.{
   DataProvision,
   RegistrationFailedMessage,
   RegistrationSuccessfulMessage,
