+++ src/test/scala/edu/ie3/simona/agent/grid/ReceivedValuesStoreSpec.scala
@@ -7,7 +7,7 @@
 package edu.ie3.simona.agent.grid
 
 import edu.ie3.datamodel.graph.SubGridGate
-import edu.ie3.simona.agent.participant2.ParticipantAgent
+import edu.ie3.simona.agent.participant.ParticipantAgent
 import edu.ie3.simona.test.common.UnitSpec
 import edu.ie3.simona.test.common.model.grid.SubGridGateMokka
 import org.apache.pekko.actor.testkit.typed.scaladsl.{
