+++ src/test/scala/edu/ie3/simona/agent/participant/ParticipantAgentMockFactory.scala
@@ -4,9 +4,9 @@
  * Research group Distribution grid planning and operation
  */
 
-package edu.ie3.simona.agent.participant2
+package edu.ie3.simona.agent.participant
 
-import edu.ie3.simona.agent.participant2.ParticipantAgent.{
+import edu.ie3.simona.agent.participant.ParticipantAgent.{
   Flex,
   FlexControlledData,
   ParticipantActivation,
