+++ src/test/scala/edu/ie3/simona/agent/participant/ParticipantAgentSpec.scala
@@ -4,7 +4,7 @@
  * Research group Distribution grid planning and operation
  */
 
-package edu.ie3.simona.agent.participant2
+package edu.ie3.simona.agent.participant
 
 import edu.ie3.datamodel.models.OperationTime
 import edu.ie3.datamodel.models.result.system.FlexOptionsResult
@@ -13,13 +13,13 @@ import edu.ie3.simona.agent.grid.GridAgentMessages.{
   AssetPowerChangedMessage,
   AssetPowerUnchangedMessage,
 }
-import edu.ie3.simona.agent.participant2.MockParticipantModel.{
+import edu.ie3.simona.agent.participant.MockParticipantModel.{
   MockRequestMessage,
   MockResponseMessage,
   MockResult,
   MockSecondaryData,
 }
-import edu.ie3.simona.agent.participant2.ParticipantAgent.{
+import edu.ie3.simona.agent.participant.ParticipantAgent.{
   DataProvision,
   GridSimulationFinished,
   NoDataProvision,
