+++ src/test/scala/edu/ie3/simona/service/primary/PrimaryServiceWorkerSqlIT.scala
@@ -9,7 +9,7 @@ package edu.ie3.simona.service.primary
 import com.dimafeng.testcontainers.{ForAllTestContainer, PostgreSQLContainer}
 import edu.ie3.datamodel.io.naming.DatabaseNamingStrategy
 import edu.ie3.datamodel.models.value.{HeatAndSValue, PValue}
-import edu.ie3.simona.agent.participant2.ParticipantAgent.{
+import edu.ie3.simona.agent.participant.ParticipantAgent.{
   DataProvision,
   PrimaryRegistrationSuccessfulMessage,
 }
