+++ src/test/scala/edu/ie3/simona/service/weather/WeatherServiceSpec.scala
@@ -8,7 +8,7 @@ package edu.ie3.simona.service.weather
 
 import com.typesafe.config.ConfigFactory
 import com.typesafe.scalalogging.LazyLogging
-import edu.ie3.simona.agent.participant2.ParticipantAgent.{
+import edu.ie3.simona.agent.participant.ParticipantAgent.{
   DataProvision,
   RegistrationFailedMessage,
   RegistrationSuccessfulMessage,
